pyproject.toml
[project]
name = "neo4j-livelihood-support"
version = "1.4.0"
description = "生活保護受給者尊厳支援データベース - Manifesto Version 1.4準拠（金銭的安全と多機関連携対応）"
readme = "docs/Manifesto_LivelihoodSupport_Graph.md"
requires-python = ">=3.11"
dependencies = [
    # コアライブラリ
    "neo4j>=5.26.0",
    "python-dotenv>=1.0.1",

    # AI関連
    "agno>=1.0.0",
    "google-genai>=1.0.0",  # agnoのGemini連携に必要

    # Streamlit UI
    "streamlit>=1.40.0",

    # MCP サーバー
    "mcp[cli]>=1.0.0",

    # オプション: ファイル読み込み用
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "pypdf>=5.1.0",

    # 認証 (Phase 1 セキュリティ基盤)
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",

    # FastAPI REST API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Monitoring
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.scripts]
# Streamlit アプリ起動
start = "streamlit run app_case_record.py"
# MCP サーバー起動
mcp-server = "python -m mcp.server"
# FastAPI サーバー起動
api-server = "uvicorn api.main:app --reload --host 0.0.0.0 --port 8000"

[tool.uv]
dev-dependencies = [
    # Testing
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",

    # Linting & Formatting
    "ruff>=0.8.0",

    # Type Checking
    "mypy>=1.13.0",

    # Security
    "bandit>=1.7.0",
    "safety>=3.0.0",

    # E2E Testing
    "playwright>=1.49.0",
    "pytest-playwright>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# =============================================================================
# Ruff (Linter & Formatter)
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "neo4j_data",
    "neo4j_logs",
    "neo4j_plugins",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # use of assert (OK in tests)
    "B008",   # function calls in argument defaults (FastAPI uses this)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106"]  # Allow asserts and hardcoded passwords in tests

# =============================================================================
# MyPy (Type Checker)
# =============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    ".venv",
    "tests",
    "neo4j_data",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# =============================================================================
# Bandit (Security Scanner)
# =============================================================================
[tool.bandit]
exclude_dirs = [".venv", "tests"]
skips = ["B101"]  # Skip assert warnings

# =============================================================================
# Coverage
# =============================================================================
[tool.coverage.run]
source = ["lib", "api"]
omit = ["tests/*", ".venv/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 60
